完善订单确认

+++ services/dou/src/main/java/cn/qihangerp/module/open/dou/domain/DouOrder.java
@@ -395,7 +395,7 @@ public class DouOrder implements Serializable {
     /**
      * 系统店铺id
      */
-    private Integer sShopId;
+    private Long sShopId;
 
     /**
      * 第一次拉取时间
